Move the Patch class to lib/
